+++ src/classes/babies/PuzzleBaby.ts
@@ -9,9 +9,10 @@ import { Baby } from "../Baby";
 
 /** D6 effect on hit. */
 export class PuzzleBaby extends Baby {
+  /** There are no items on the Cathedral. */
   override isValid(): boolean {
     const effectiveStage = getEffectiveStage();
-    return effectiveStage !== (LevelStage.SHEOL_CATHEDRAL as int);
+    return effectiveStage !== LevelStage.SHEOL_CATHEDRAL;
   }
 
   @CallbackCustom(ModCallbackCustom.ENTITY_TAKE_DMG_PLAYER)
