+++ src/classes/babies/TwinBaby.ts
@@ -7,6 +7,6 @@ export class TwinBaby extends Baby {
   /** If they mess up and go past the Boss Room on Womb 2, they can get the wrong path. */
   override isValid(): boolean {
     const effectiveStage = getEffectiveStage();
-    return effectiveStage !== (LevelStage.WOMB_2 as int);
+    return effectiveStage !== LevelStage.WOMB_2;
   }
 }
